requirements.txt
Appium-Python-Client>=3.0.0
pytest>=7.0.0
pytest-html>=3.2.0
